pyproject.toml
[project]
name = "tt-smi"
version = "1.0.0"
description = "ncurses based hardware monitoring for Tenstorrent silicon"
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
authors = [
  { name = "Sam Bansal", email = "sbansal@tenstorrent.com" }
]
maintainers = [
  { name = "Sam Bansal", email = "sbansal@tenstorrent.com" }
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console :: Curses",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
  'art',
  'bitstring',
  'blessed',
  'cryptography',
  'distro',
  'elasticsearch==7.17.4',
  'elasticsearch_dsl',
  'fabric',
  'glob2',
  'jsons',
  'numpy',
  'openpyxl==3.0.10',
  'pandas',
  'prometheus_client',
  'psutil',
  'pydantic==1.*',
  'pyfiglet',
  'pyinstaller',
  'pyluwen @ file:///home/sbansal/luwen/crates/pyluwen',
  'pyserial',
  'python-constraint',
  'pyyaml',
  'requests',
  'rich',
  'ruamel.yaml',
  'smbus2==0.4.2',
  'tabulate',
  'termcolor',
  'textual-dev',
  'tqdm',
  'wheel',
  'xlsxwriter',
]

[project.urls]
"Homepage" = "http://tenstorrent.com"
"Bug Reports" = "https://github.com/tenstorrent/tt-smi/issues"
"Source" = "https://github.com/tenstorrent/tt-smi"

[project.scripts]
tt-smi = "tt_smi:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"*" = [
  'data/version.txt',
  "*.css",
  "data/*/*.yaml",
  "data/*/*/*.yaml",
]

[tools.setuptools.exclude-package-data]
"*" = [
  'build',
]

[tool.setuptools.packages.find]
where = [ ".", ]

[build-system]
# These are the assumed default build requirements from pip:
# https://pip.pypa.io/en/stable/reference/pip/#pep-517-and-518-support
requires = [
  "setuptools>=43.0.0",
  "wheel"
]
build-backend = "setuptools.build_meta"

# vim: expandtab:ts=2:sw=2:ai
